Fix installing puppet in Debian.
This patch fixes the install of the puppet packages in Debian.

+++ all-in-one.sh
@@ -48,9 +48,13 @@ if is_fedora; then
     DASHBOARD="dashboard"
 elif uses_debs; then
     print_header 'Setup (Debian based)'
-    sudo apt-get remove -y --purge facter puppet puppet-common
     sudo apt-get update
     sudo apt-get install -y libxml2-dev libxslt-dev zlib1g-dev ruby wget lsb-release curl apt-transport-https adduser
+    # Only purge puppet in Ubuntu, in Debian we use
+    # what the distribution provides.
+    if [ $(lsb_release --id -s) = "Ubuntu" ] ; then
+        sudo apt-get remove -y --purge facter puppet puppet-common
+    fi
     DASHBOARD="horizon"
 fi
